Separate classes for bookmark and episode actions
